requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
dotmap==1.3.30
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.14.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
yarl==1.9.4
